+++ tests/attributes/missingness/modules/uds/test_missingness_a3.py
@@ -94,3 +94,63 @@ class TestUDSFormA3V4Missingness:
 
         attr = UDSFormA3V4Missingness(uds_table)
         assert attr._missingness_kid2ago() == INFORMED_MISSINGNESS
+
+    def test_prev_sibs_kids(self, uds_table):
+        """Test SIBS/KIDS provided at previous vist."""
+        # from prev visit
+        uds_table.update(
+            {
+                "_prev_record": {
+                    "info": {
+                        "forms": {"json": {"visitdate": "2020-01-01"}},
+                        "resolved": {"sibs": 2, "kids": "5"},
+                    }
+                }
+            }
+        )
+        uds_table["file.info.forms.json"].update(
+            {
+                "formver": 4,
+                "packet": "F",
+                "nwinfsib": 0,
+                "nwinfkid": "1",
+                "sibs": 66,
+                "kids": "66",
+            }
+        )
+
+        attr = UDSFormA3V4Missingness(uds_table)
+        assert attr._missingness_sibs() == 2
+        assert attr._missingness_kids() == 5
+
+    def test_sibs_kids(self, uds_table):
+        """Test sibs/kids when not a previous visit."""
+        attr = UDSFormA3V4Missingness(uds_table)
+        assert attr._missingness_sibs() == INFORMED_MISSINGNESS
+        assert attr._missingness_kids() == INFORMED_MISSINGNESS
+
+        uds_table["file.info.forms.json"].update(
+            {
+                "packet": "I4",
+                "sibs": 66,  # technically not allowed but will go through on curation
+                "kids": "3",
+            }
+        )
+
+        attr = UDSFormA3V4Missingness(uds_table)
+        assert attr._missingness_sibs() == 66
+        assert attr._missingness_kids() == 3
+
+        # not V4
+        uds_table["file.info.forms.json"].update(
+            {
+                "formver": 3,
+                "packet": "I",
+                "sibs": None,
+                "kids": "7",
+            }
+        )
+
+        attr = UDSFormA3V4Missingness(uds_table)
+        assert attr._missingness_sibs() == INFORMED_MISSINGNESS
+        assert attr._missingness_kids() == 7